requirements.txt
apache-airflow~=2.10.2
apache-airflow-providers-amazon
